create static method Degree.add, and update functions to use it
