+++ packages/database/src/entity/Degree.ts
@@ -1,5 +1,8 @@
 import { Entity, PrimaryGeneratedColumn, Column, ManyToMany, JoinTable } from 'typeorm'
 import { Module } from './Module'
+import { container, AppDataSource } from '../data-source'
+import { DegreeProps } from '../../types/modtree'
+import { In } from 'typeorm'
 
 @Entity({ name: 'degree' })
 export class Degree {
@@ -12,4 +15,24 @@ export class Degree {
 
   @Column()
   title: string
+
+  /**
+   * Adds a Degree to DB
+   */
+  static async add(props: DegreeProps): Promise<void> {
+    await container(async() => {
+      // find modules required, to create many-to-many relation
+      const repo = AppDataSource.getRepository(Module)
+      const modulesRequired = await repo.find({
+        where: {
+          moduleCode: In(props.moduleCodes),
+        },
+      })
+
+      const degree = new Degree()
+      degree.title = props.title
+      degree.modulesRequired = modulesRequired
+      await AppDataSource.manager.save(degree)
+    })
+  }
 }
