+++ packages/database/types/modtree.d.ts
@@ -9,3 +9,8 @@ export type UserProps = {
   graduationYear: number,
   graduationSemester: number,
 }
+
+export type DegreeProps = {
+  moduleCodes: string[]
+  title: string
+}
